feat: Remove Electron back compat. (v5)
