--- src/common/index.ts
@@ -1,5 +1,0 @@
-export * from './ipc';
-export * from './normalize';
-export * from './merge';
-export * from './mode';
-export * from './process';
