+++ src/common/ipc.ts
@@ -1,5 +1,20 @@
 import { MeasurementUnit, Primitive } from '@sentry/types';
 
+/** Ways to communicate between the renderer and main process  */
+export enum IPCMode {
+  /** Configures Electron IPC to receive messages from renderers */
+  Classic = 1,
+  /** Configures Electron protocol module to receive messages from renderers */
+  Protocol = 2,
+  /**
+   * Configures both methods for best compatibility.
+   *
+   * Renderers favour IPC but fall back to protocol if IPC has not
+   * been configured in a preload script
+   */
+  Both = 3,
+}
+
 export const PROTOCOL_SCHEME = 'sentry-ipc';
 
 export enum IPCChannel {
