+++ src/main/merge.ts
