+++ src/main/normalize.ts
