+++ test/e2e/test-apps/anr/anr-renderer/recipe.yml
@@ -1,4 +1,3 @@
 description: ANR Renderer Event
 category: ANR
 command: yarn
-condition: version.major >= 15
