+++ test/e2e/test-apps/javascript/preload-error-protocol/recipe.yml
@@ -1,4 +1,3 @@
 description: JavaScript Preload Error via Protocol
 category: JavaScript
 command: yarn
-condition: version.major >= 5
