+++ test/e2e/test-apps/native-electron/main-custom-release/recipe.yml
@@ -1,4 +1,3 @@
 description: Native Main Crash (Custom Release Name)
 category: Native (Electron Uploader)
 command: yarn
-condition: version.major >= 9
