+++ test/e2e/test-apps/native-electron/renderer-custom-release/recipe.yml
@@ -1,4 +1,3 @@
 description: Native Renderer Crash (Custom Release Name)
 category: Native (Electron Uploader)
 command: yarn
-condition: version.major >= 9
