+++ test/e2e/test-apps/native-sentry/gpu/recipe.yml
@@ -2,5 +2,5 @@ description: Native GPU Crash
 category: Native (Sentry Uploader)
 # We will only get GPU crashes from Crashpad
 # and the chrome://gpucrash was only added in v13
-condition: usesCrashpad && version.major >= 13
+condition: usesCrashpad
 command: yarn
