+++ test/e2e/test-apps/native-sentry/renderer-force-crash/recipe.yml
@@ -1,4 +1,4 @@
 description: Native Renderer forcefullyCrashRenderer
 category: Native (Sentry Uploader)
 command: yarn
-condition: usesCrashpad && version.major >= 12
+condition: usesCrashpad
