+++ test/e2e/test-apps/other/browser-profiling/recipe.yml
@@ -1,4 +1,3 @@
 description: Browser Profiling
 command: yarn
-# Browser Tracing fails on Electron v24
-condition: version.major >= 15
+
