+++ test/e2e/test-apps/other/error-after-ready/recipe.yml
@@ -1,4 +1,3 @@
 description: Error thrown if init after app ready
 command: yarn
-condition: version.major >= 5
 expectedError: should be initialized before the Electron app 'ready'
