+++ test/e2e/test-apps/other/error-entry-point/recipe.yml
@@ -1,6 +1,5 @@
 description: Error thrown if renderer code loaded in main process
 command: yarn && yarn build
-condition: version.major >= 15
 timeout: 120
 expectedError: This code is intended to run in the Electron renderer
 skipEsmAutoTransform: true
