+++ test/e2e/test-apps/sessions/abnormal-exit-electron-uploader/recipe.yml
@@ -1,5 +1,4 @@
 description: Session with Abnormal Exit (Electron Uploader)
 category: Sessions
 command: yarn
-condition: version.major >= 9
 runTwice: true
