+++ test/e2e/test-apps/sessions/native-crash-main-electron-uploader/recipe.yml
@@ -1,5 +1,5 @@
 description: Session with Native Main Crash (Electron Uploader)
 category: Sessions
 command: yarn
-condition: version.major >= 9 && usesCrashpad
+condition: usesCrashpad
 runTwice: true
