+++ test/e2e/test-apps/sessions/native-crash-renderer-electron-uploader/recipe.yml
@@ -1,4 +1,3 @@
 description: Session with Native Renderer Crash (Electron Uploader)
 category: Sessions
 command: yarn
-condition: version.major >= 9
