+++ test/e2e/test-apps/sessions/window-bad/recipe.yml
@@ -1,4 +1,3 @@
 description: Bad Session - Window
 category: Sessions
 command: yarn
-condition: version.major >= 12
