+++ test/e2e/test-apps/sessions/window-good/recipe.yml
@@ -1,4 +1,3 @@
 description: Good Session - Window
 category: Sessions
 command: yarn
-condition: version.major >= 12
